requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
get==2019.4.13
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
post==2019.4.13
public==2019.4.13
python-chess==0.29.0
query-string==2019.4.13
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
